requirements.txt
streamlit
PyPDF2
langchain
langchain-openai
openai
python-dotenv
faiss-cpu
sentence-transformers
gtts
pandas
numpy